+++ packages/core/src/api/events/initialize.ts
@@ -1,4 +1,5 @@
 import type { Config } from '../../config';
+import { getInternalMetas } from '../../internal';
 import type { InternalLogger } from '../../internalLogger';
 import { captureMetas, type Metas } from '../../metas';
 import { TransportItemType } from '../../transports';
@@ -37,6 +38,9 @@ export function initializeEventsAPI({
     { skipDedupe, spanContext, timestampOverwriteMs, customPayloadTransformer = (payload: EventEvent) => payload } = {}
   ) => {
     try {
+      if (!getInternalMetas(metas).shouldCapture()) {
+        return;
+      }
       const attrs = stringifyObjectValues(attributes);
 
       const payload = customPayloadTransformer({